# Speed up with MutableArithmetics