--- a/src/comp.jl
+++ b/src/comp.jl
@@ -1,14 +1,14 @@
 isapproxzero(x::T; kws...) where {T<:Real} = x == zero(T)
-isapproxzero(x::T; ztol=Base.rtoldefault(T)) where {T<:AbstractFloat} = abs(x) < ztol
-isapproxzero(x::AbstractVector{T}; kws...) where {T<:Real} = isapproxzero(maximum(abs.(x)); kws...)
+isapproxzero(x::T; tol=Base.rtoldefault(T)) where {T<:AbstractFloat} = abs(x) < tol
+isapproxzero(x::AbstractVector{T}; kws...) where {T<:Real} = isapproxzero(maximum(abs, x); kws...)
 #isapproxzero(x::Union{SymPoint, Ray, Line}; kws...) = isapproxzero(coord(x); kws...)
 isapproxzero(x::Union{Ray, Line}; kws...) = isapproxzero(coord(x); kws...)
 isapproxzero(h::HRepElement; kws...) = isapproxzero(h.a; kws...) && isapproxzero(h.β; kws...)
 
-_isapprox(x::Union{T, AbstractArray{T}}, y::Union{T, AbstractArray{T}}) where {T<:Union{Integer, Rational}} = x == y
+_isapprox(x::Union{T, AbstractArray{T}}, y::Union{T, AbstractArray{T}}; tol) where {T<:Union{Integer, Rational}} = x == y
 # I check with zero because isapprox(0, 1e-100) is false...
 # but isapprox(1e-100, 2e-100) should be false
-_isapprox(x, y) = (isapproxzero(x) ? isapproxzero(y) : (isapproxzero(y) ? isapproxzero(x) : isapprox(x, y)))
+_isapprox(x, y; tol) = (isapproxzero(x; tol) ? isapproxzero(y; tol) : (isapproxzero(y; tol) ? isapproxzero(x; tol) : isapprox(x, y; rtol = tol)))
 
 #Base.isapprox(r::SymPoint, s::SymPoint) = _isapprox(coord(r), coord(s)) || _isapprox(coord(r), -coord(s))
 function _scaleray(r::Union{Line, Ray}, s::Union{Line, Ray})
@@ -46,15 +46,15 @@ function Base.isapprox(h1::HalfSpace, h2::HalfSpace)
     _isapprox(a1, a2) && _isapprox(β1, β2)
 end
 
-_lt(x::T, y::T) where {T<:Real} = x < y
-_lt(x::T, y::T) where {T<:AbstractFloat} = x < y && !_isapprox(x, y)
-_lt(x::S, y::T) where {S<:Real,T<:Real} = _lt(promote(x, y)...)
-_gt(x::S, y::T) where {S<:Real, T<:Real} = _lt(y, x)
-_leq(x::T, y::T) where {T<:Real} = x <= y
-_leq(x::T, y::T) where {T<:AbstractFloat} = x <= y || _isapprox(x, y)
-_leq(x::S, y::T) where {S<:Real,T<:Real} = _leq(promote(x, y)...)
-_geq(x::T, y::T) where {T<:Real} = _leq(y, x)
-_pos(x::T) where {T<:Real} = _gt(x, zero(T))
-_neg(x::T) where {T<:Real} = _lt(x, zero(T))
-_nonneg(x::T) where {T<:Real} = _geq(x, zero(T))
-_nonpos(x::T) where {T<:Real} = _leq(x, zero(T))
+_lt(x::T, y::T; tol) where {T<:Real} = x < y
+_lt(x::T, y::T; tol) where {T<:AbstractFloat} = x < y && !_isapprox(x, y; tol)
+_lt(x::S, y::T; tol) where {S<:Real,T<:Real} = _lt(promote(x, y)...; tol)
+_gt(x::S, y::T; tol) where {S<:Real, T<:Real} = _lt(y, x; tol)
+_leq(x::T, y::T; tol) where {T<:Real} = x <= y
+_leq(x::T, y::T; tol) where {T<:AbstractFloat} = x <= y || _isapprox(x, y; tol)
+_leq(x::S, y::T; tol) where {S<:Real,T<:Real} = _leq(promote(x, y)...; tol)
+_geq(x::T, y::T; tol) where {T<:Real} = _leq(y, x; tol)
+_pos(x::T; tol) where {T<:Real} = _gt(x, zero(T); tol)
+_neg(x::T; tol) where {T<:Real} = _lt(x, zero(T); tol)
+_nonneg(x::T; tol) where {T<:Real} = _geq(x, zero(T); tol)
+_nonpos(x::T; tol) where {T<:Real} = _leq(x, zero(T); tol)