… .github/workflows/ci.yml
… Project.toml
… docs/src/optimization.md
… src/interval.jl
… src/linearity.jl
… src/planar.jl
… src/polyhedron.jl
… src/redundancy.jl
… src/repelemop.jl
… src/triangulation.jl
… src/vrep_optimizer.jl
… test/issue301.jl
… test/lp_to_polyhedra.jl
… test/lphrep.jl
… test/redundancy.jl